+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/ads/AdMobDisplayer.kt
@@ -25,7 +25,8 @@ class AdMobDisplayer(private val context: Activity) : MiniAppAdDisplayer, Corout
     override val coroutineContext: CoroutineContext
         get() = Dispatchers.Main
     private val interstitialAdMap = HashMap<String, InterstitialAd>()
-    private val rewardedAdMap = HashMap<String, RewardedAd>()
+    @VisibleForTesting(otherwise = VisibleForTesting.NONE)
+    internal val rewardedAdMap = HashMap<String, RewardedAd>()
 
     @VisibleForTesting
     internal fun initAdMap(
@@ -55,6 +56,7 @@ class AdMobDisplayer(private val context: Activity) : MiniAppAdDisplayer, Corout
                     }
 
                     override fun onAdFailedToLoad(adError: LoadAdError) {
+                        interstitialAdMap.remove(adUnitId)
                         onFailed.invoke(adError.message)
                     }
                 }
@@ -92,7 +94,7 @@ class AdMobDisplayer(private val context: Activity) : MiniAppAdDisplayer, Corout
                 val ad = RewardedAd(context, adUnitId)
                 rewardedAdMap[adUnitId] = ad
 
-                val adLoadCallback = createRewardedAdLoadCallback(onLoaded, onFailed)
+                val adLoadCallback = createRewardedAdLoadCallback(adUnitId, onLoaded, onFailed)
 
                 ad.loadAd(AdRequest.Builder().build(), adLoadCallback)
             }
@@ -118,6 +120,7 @@ class AdMobDisplayer(private val context: Activity) : MiniAppAdDisplayer, Corout
 
     @VisibleForTesting
     internal fun createRewardedAdLoadCallback(
+        adUnitId: String,
         onLoaded: () -> Unit,
         onFailed: (String) -> Unit
     ): RewardedAdLoadCallback = object : RewardedAdLoadCallback() {
@@ -127,6 +130,7 @@ class AdMobDisplayer(private val context: Activity) : MiniAppAdDisplayer, Corout
         }
 
         override fun onRewardedAdFailedToLoad(adError: LoadAdError) {
+            rewardedAdMap.remove(adUnitId)
             onFailed.invoke(adError.message)
         }
     }