+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/display/MiniAppWebviewSpec.kt
@@ -4,7 +4,6 @@ import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
-import android.util.AndroidRuntimeException
 import android.view.View
 import android.view.ViewGroup
 import android.webkit.GeolocationPermissions
@@ -242,10 +241,14 @@ class MiniAppWebviewSpec : BaseWebViewSpec() {
     @Test
     fun `should send response with escaped backtick characters`() {
         val spyMiniAppWebView = spy(miniAppWebView)
-        spyMiniAppWebView.runSuccessCallback("test_id", "`test response`")
 
+        spyMiniAppWebView.runSuccessCallback("test_id", "`test response`")
         Verify on spyMiniAppWebView that spyMiniAppWebView.evaluateJavascript(
             argWhere { it.contains("""`\`test response\``""") }, any())
+
+        spyMiniAppWebView.runErrorCallback("test_id", "`error response`")
+        Verify on spyMiniAppWebView that spyMiniAppWebView.evaluateJavascript(
+            argWhere { it.contains("""`\`error response\``""") }, any())
     }
 }